test(ComputerTest): 힌트 반환 기능 테스트 구현
