+++ docs/README.md
@@ -40,7 +40,7 @@
     - [X] `예외사항` : 세 자리가 아니라면 예외 발생
     - [X] `예외사항` : 1~9사이의 숫자가 아니라면 예외 발생
     - [X] `예외사항` : 중복된 숫자가 있다면 예외 발생
-        - [ ] `IllegalArgumentException`을 발생하고 프로그램 종료
+        - [x] `IllegalArgumentException`을 발생하고 프로그램 종료
 - [X] 사용자로부터 게임 재시작 여부를 입력받는다.
     - `1` : 게임 재시작
     - `2` : 게임 종료
@@ -61,10 +61,9 @@
 
 ### 힌트 반환 기능 테스트 ###
 
-- [ ] 사용자 입력에 대한 힌트를 반환한다.
-    - [ ] 같은 수가 같은 자리에 있으면 `스트라이크`
-    - [ ] 같은 수가 다른 자리에 있으면 `볼`
-    - [ ] 같은 수가 전혀 없으면 `낫싱`
+- [X] 사용자 입력에 대한 힌트를 반환한다.
+    - [X] 같은 수가 같은 자리에 있으면 `스트라이크`
+    - [X] 같은 수가 다른 자리에 있으면 `볼`
 
 ---
